+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* [3461](https://github.com/zeta-chain/node/pull/3461) - enable multiple confirmation count values in chain params
+* [3461](https://github.com/zeta-chain/node/pull/3461) - add new 'ConfirmationParams' field to chain params to enable multiple confirmation count values, deprecating `confirmation_count`
 
 ### Tests